Fix bug in first operand's variable type
The result of operations was being returned as a Number instead of a
String. This meant that, when the result was a floating point number,
it wasn't possible to use the function String.includes() to check if
the value already contained a "." character. This was fixed to preserve
the consistency of interactions with the calculator.